src/dashboard/requirements.txt
# Dashboard dependencies
streamlit>=1.28.0
plotly>=5.15.0
pandas>=2.0.0
numpy>=1.24.0
psycopg2-binary>=2.9.0
redis>=4.5.0
psutil>=5.9.0

# Data processing
requests>=2.31.0

# Optional for enhanced visualizations
altair>=5.0.0
pillow>=10.0.0
